frontend/requirements.txt
# Core Libraries
Flask
gunicorn

# Google Cloud Libraries
google-cloud-storage
google-cloud-tasks
google-cloud-firestore
google-auth
requests
google-cloud-logging
